% Auto-generated: do not edit by hand
\name{imgPanZoom}

\alias{imgPanZoom}

\title{ImgPanZoom component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
imgPanZoom(id=NULL, height=NULL, width=NULL, src=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{height}{Numeric. The height is used to define the height of the canvas}

\item{width}{Numeric. The width is used to define the width of the canvas}

\item{src}{Character. The src is the image url to be pasted in the canvas}
}
